+++ salt/daemons/masterapi.py
@@ -15,7 +15,7 @@ import stat
 
 # Import salt libs
 import salt.crypt
-import salt.utils
+import salt.utils  # Can be removed once check_whitelist_blacklist, expr_match, get_values_of_matching_keys are moved
 import salt.cache
 import salt.client
 import salt.payload
@@ -29,6 +29,7 @@ import salt.key
 import salt.fileserver
 import salt.utils.args
 import salt.utils.atomicfile
+import salt.utils.dictupdate
 import salt.utils.event
 import salt.utils.files
 import salt.utils.gitfs
@@ -38,6 +39,7 @@ import salt.utils.gzip_util
 import salt.utils.jid
 import salt.utils.minions
 import salt.utils.platform
+import salt.utils.user
 import salt.utils.verify
 from salt.defaults import DEFAULT_TARGET_DELIM
 from salt.pillar import git_pillar
@@ -227,7 +229,7 @@ def access_keys(opts):
     acl_users = set(publisher_acl.keys())
     if opts.get('user'):
         acl_users.add(opts['user'])
-    acl_users.add(salt.utils.get_user())
+    acl_users.add(salt.utils.user.get_user())
     for user in acl_users:
         log.info('Preparing the %s key for local communication', user)
         key = mk_key(opts, user)
@@ -286,7 +288,7 @@ class AutoKey(object):
             pwnam = pwd.getpwnam(user)
             uid = pwnam[2]
             gid = pwnam[3]
-            groups = salt.utils.get_gid_list(user, include_default=False)
+            groups = salt.utils.user.get_gid_list(user, include_default=False)
         except KeyError:
             log.error(
                 'Failed to determine groups for user {0}. The user is not '
