+++ salt/modules/mac_user.py
@@ -27,6 +27,7 @@ import salt.utils
 import salt.utils.args
 import salt.utils.decorators.path
 import salt.utils.stringutils
+import salt.utils.user
 from salt.utils.locales import sdecode as _sdecode
 from salt.exceptions import CommandExecutionError, SaltInvocationError
 
@@ -454,7 +455,7 @@ def list_groups(name):
 
         salt '*' user.list_groups foo
     '''
-    groups = [group for group in salt.utils.get_group_list(name)]
+    groups = [group for group in salt.utils.user.get_group_list(name)]
     return groups
