+++ salt/modules/pw_user.py
@@ -47,9 +47,10 @@ except ImportError:
 from salt.ext import six
 
 # Import salt libs
-import salt.utils
+import salt.utils  # Can be removed once is_true is moved
 import salt.utils.args
 import salt.utils.locales
+import salt.utils.user
 from salt.exceptions import CommandExecutionError
 
 log = logging.getLogger(__name__)
@@ -489,7 +490,7 @@ def list_groups(name):
 
         salt '*' user.list_groups foo
     '''
-    return salt.utils.get_group_list(name)
+    return salt.utils.user.get_group_list(name)
 
 
 def list_users():
