+++ salt/modules/rabbitmq.py
@@ -16,10 +16,10 @@ import random
 import string
 
 # Import salt libs
-import salt.utils
 import salt.utils.itertools
 import salt.utils.path
 import salt.utils.platform
+import salt.utils.user
 from salt.ext import six
 from salt.exceptions import SaltInvocationError
 from salt.ext.six.moves import range
@@ -222,7 +222,7 @@ def list_users(runas=None):
     # Due to this, don't use a default value for
     # runas in Windows.
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'list_users', '-q'],
         runas=runas,
@@ -245,7 +245,7 @@ def list_vhosts(runas=None):
         salt '*' rabbitmq.list_vhosts
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'list_vhosts', '-q'],
         runas=runas,
@@ -265,7 +265,7 @@ def user_exists(name, runas=None):
         salt '*' rabbitmq.user_exists rabbit_user
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     return name in list_users(runas=runas)
 
 
@@ -280,7 +280,7 @@ def vhost_exists(name, runas=None):
         salt '*' rabbitmq.vhost_exists rabbit_host
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     return name in list_vhosts(runas=runas)
 
 
@@ -303,7 +303,7 @@ def add_user(name, password=None, runas=None):
             string.ascii_uppercase + string.digits) for x in range(15))
 
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
 
     if salt.utils.platform.is_windows():
         # On Windows, if the password contains a special character
@@ -351,7 +351,7 @@ def delete_user(name, runas=None):
         salt '*' rabbitmq.delete_user rabbit_user
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'delete_user', name],
         python_shell=False,
@@ -372,7 +372,7 @@ def change_password(name, password, runas=None):
         salt '*' rabbitmq.change_password rabbit_user password
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     if salt.utils.platform.is_windows():
         # On Windows, if the password contains a special character
         # such as '|', normal execution will fail. For example:
@@ -408,7 +408,7 @@ def clear_password(name, runas=None):
         salt '*' rabbitmq.clear_password rabbit_user
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'clear_password', name],
         runas=runas,
@@ -433,7 +433,7 @@ def check_password(name, password, runas=None):
     # try to get the rabbitmq-version - adapted from _get_rabbitmq_plugin
 
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
 
     try:
         res = __salt__['cmd.run']([RABBITMQCTL, 'status'], runas=runas, python_shell=False)
@@ -508,7 +508,7 @@ def add_vhost(vhost, runas=None):
         salt '*' rabbitmq add_vhost '<vhost_name>'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'add_vhost', vhost],
         runas=runas,
@@ -529,7 +529,7 @@ def delete_vhost(vhost, runas=None):
         salt '*' rabbitmq.delete_vhost '<vhost_name>'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'delete_vhost', vhost],
         runas=runas,
@@ -549,7 +549,7 @@ def set_permissions(vhost, user, conf='.*', write='.*', read='.*', runas=None):
         salt '*' rabbitmq.set_permissions 'myvhost' 'myuser'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'set_permissions', '-p',
          vhost, user, conf, write, read],
@@ -570,7 +570,7 @@ def list_permissions(vhost, runas=None):
         salt '*' rabbitmq.list_permissions '/myvhost'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'list_permissions', '-q', '-p', vhost],
         runas=runas,
@@ -590,7 +590,7 @@ def list_user_permissions(name, runas=None):
         salt '*' rabbitmq.list_user_permissions 'user'.
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'list_user_permissions', name, '-q'],
         runas=runas,
@@ -609,7 +609,7 @@ def set_user_tags(name, tags, runas=None):
         salt '*' rabbitmq.set_user_tags 'myadmin' 'administrator'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
 
     if not isinstance(tags, (list, tuple)):
         tags = [tags]
@@ -633,7 +633,7 @@ def status(runas=None):
         salt '*' rabbitmq.status
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'status'],
         runas=runas,
@@ -653,7 +653,7 @@ def cluster_status(runas=None):
         salt '*' rabbitmq.cluster_status
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'cluster_status'],
         runas=runas,
@@ -678,7 +678,7 @@ def join_cluster(host, user='rabbit', ram_node=None, runas=None):
     cmd.append('{0}@{1}'.format(user, host))
 
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     stop_app(runas)
     res = __salt__['cmd.run_all'](cmd, runas=runas, python_shell=False)
     start_app(runas)
@@ -697,7 +697,7 @@ def stop_app(runas=None):
         salt '*' rabbitmq.stop_app
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'stop_app'],
         runas=runas,
@@ -717,7 +717,7 @@ def start_app(runas=None):
         salt '*' rabbitmq.start_app
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'start_app'],
         runas=runas,
@@ -737,7 +737,7 @@ def reset(runas=None):
         salt '*' rabbitmq.reset
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'reset'],
         runas=runas,
@@ -757,7 +757,7 @@ def force_reset(runas=None):
         salt '*' rabbitmq.force_reset
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'force_reset'],
         runas=runas,
@@ -777,7 +777,7 @@ def list_queues(runas=None, *args):
         salt '*' rabbitmq.list_queues messages consumers
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [RABBITMQCTL, 'list_queues', '-q']
     cmd.extend(args)
     res = __salt__['cmd.run_all'](cmd, runas=runas, python_shell=False)
@@ -799,7 +799,7 @@ def list_queues_vhost(vhost, runas=None, *args):
         salt '*' rabbitmq.list_queues messages consumers
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [RABBITMQCTL, 'list_queues', '-q', '-p', vhost]
     cmd.extend(args)
     res = __salt__['cmd.run_all'](cmd, runas=runas, python_shell=False)
@@ -822,7 +822,7 @@ def list_policies(vhost="/", runas=None):
     '''
     ret = {}
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'list_policies', '-q', '-p', vhost],
         runas=runas,
@@ -864,7 +864,7 @@ def set_policy(vhost, name, pattern, definition, priority=None, apply_to=None, r
         salt '*' rabbitmq.set_policy / HA '.*' '{"ha-mode":"all"}'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     if isinstance(definition, dict):
         definition = json.dumps(definition)
     if not isinstance(definition, six.string_types):
@@ -895,7 +895,7 @@ def delete_policy(vhost, name, runas=None):
         salt '*' rabbitmq.delete_policy / HA'
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     res = __salt__['cmd.run_all'](
         [RABBITMQCTL, 'clear_policy', '-p', vhost, name],
         runas=runas,
@@ -917,7 +917,7 @@ def policy_exists(vhost, name, runas=None):
         salt '*' rabbitmq.policy_exists / HA
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     policies = list_policies(runas=runas)
     return bool(vhost in policies and name in policies[vhost])
 
@@ -933,7 +933,7 @@ def list_available_plugins(runas=None):
             salt '*' rabbitmq.list_available_plugins
         '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [_get_rabbitmq_plugin(), 'list', '-m']
     ret = __salt__['cmd.run_all'](cmd, python_shell=False, runas=runas)
     _check_response(ret)
@@ -951,7 +951,7 @@ def list_enabled_plugins(runas=None):
             salt '*' rabbitmq.list_enabled_plugins
         '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [_get_rabbitmq_plugin(), 'list', '-m', '-e']
     ret = __salt__['cmd.run_all'](cmd, python_shell=False, runas=runas)
     _check_response(ret)
@@ -969,7 +969,7 @@ def plugin_is_enabled(name, runas=None):
         salt '*' rabbitmq.plugin_is_enabled rabbitmq_plugin_name
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     return name in list_enabled_plugins(runas)
 
 
@@ -984,7 +984,7 @@ def enable_plugin(name, runas=None):
         salt '*' rabbitmq.enable_plugin foo
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [_get_rabbitmq_plugin(), 'enable', name]
     ret = __salt__['cmd.run_all'](cmd, runas=runas, python_shell=False)
     return _format_response(ret, 'Enabled')
@@ -1001,7 +1001,7 @@ def disable_plugin(name, runas=None):
         salt '*' rabbitmq.disable_plugin foo
     '''
     if runas is None and not salt.utils.platform.is_windows():
-        runas = salt.utils.get_user()
+        runas = salt.utils.user.get_user()
     cmd = [_get_rabbitmq_plugin(), 'disable', name]
     ret = __salt__['cmd.run_all'](cmd, runas=runas, python_shell=False)
     return _format_response(ret, 'Disabled')
