+++ salt/utils/verify.py
@@ -27,9 +27,9 @@ from salt.log.setup import LOG_LEVELS
 from salt.exceptions import SaltClientError, SaltSystemExit, \
     CommandExecutionError
 import salt.defaults.exitcodes
-import salt.utils  # Can be removed once get_jid_list and get_user are moved
 import salt.utils.files
 import salt.utils.platform
+import salt.utils.user
 
 log = logging.getLogger(__name__)
 
@@ -206,7 +206,7 @@ def verify_env(dirs, user, permissive=False, pki_dir='', skip_extra=False):
         pwnam = pwd.getpwnam(user)
         uid = pwnam[2]
         gid = pwnam[3]
-        groups = salt.utils.get_gid_list(user, include_default=False)
+        groups = salt.utils.user.get_gid_list(user, include_default=False)
 
     except KeyError:
         err = ('Failed to prepare the Salt environment for user '
@@ -302,7 +302,7 @@ def check_user(user):
     '''
     if salt.utils.platform.is_windows():
         return True
-    if user == salt.utils.get_user():
+    if user == salt.utils.user.get_user():
         return True
     import pwd  # after confirming not running Windows
     try:
@@ -311,7 +311,7 @@ def check_user(user):
             if hasattr(os, 'initgroups'):
                 os.initgroups(user, pwuser.pw_gid)  # pylint: disable=minimum-python-version
             else:
-                os.setgroups(salt.utils.get_gid_list(user, include_default=False))
+                os.setgroups(salt.utils.user.get_gid_list(user, include_default=False))
             os.setgid(pwuser.pw_gid)
             os.setuid(pwuser.pw_uid)
 
@@ -383,7 +383,7 @@ def check_path_traversal(path, user='root', skip_perm_errors=False):
             if not os.path.exists(tpath):
                 msg += ' Path does not exist.'
             else:
-                current_user = salt.utils.get_user()
+                current_user = salt.utils.user.get_user()
                 # Make the error message more intelligent based on how
                 # the user invokes salt-call or whatever other script.
                 if user != current_user:
