+++ tests/unit/modules/test_pw_user.py
@@ -316,7 +316,7 @@ class PwUserTestCase(TestCase, LoaderModuleMockMixin):
         '''
         mock_group = 'saltgroup'
 
-        with patch('salt.utils.get_group_list', MagicMock(return_value=[mock_group])):
+        with patch('salt.utils.user.get_group_list', MagicMock(return_value=[mock_group])):
             self.assertEqual(pw_user.list_groups('name'), [mock_group])
 
     def test_list_users(self):
