+++ tests/unit/modules/test_useradd.py
@@ -355,7 +355,7 @@ class UserAddTestCase(TestCase, LoaderModuleMockMixin):
         '''
         Test if it return a list of groups the named user belongs to
         '''
-        with patch('salt.utils.get_group_list', MagicMock(return_value='Salt')):
+        with patch('salt.utils.user.get_group_list', MagicMock(return_value='Salt')):
             self.assertEqual(useradd.list_groups('name'), 'Salt')
 
     # 'list_users' function tests: 1
